+++ src/js/experience.js
@@ -1,5 +1,6 @@
 import dropdown from 'bootstrap/js/dist/dropdown.js';
-import { Slider1 } from './libraries.js';
+import { Slider1 } from './slider.js';
+import { manageLogin } from './loguin.js';
 
 let myExperienceSlider;
 
@@ -8,3 +9,4 @@ if (window.visualViewport.width < 992) {
     document.querySelector('.slider-experience')
   );
 }
+manageLogin();
