+++ src/js/slider.js
